Fix exception rethrow in async functions
In the original version if the function inside `Task.Run` throws an
exception it won't get propagated.

This version fixes that by awaiting on the task,
and ensures that the context is released so you won't get
an error like this within a `using` statement:

```
SevenZipExtractor instance must not be disposed
while making an asynchronous method call.
```