Calculate UniqueID via increment
Instead of trying to find unique id randomly, which is costly, do it sequentially instead.
In this case, we have ~4 billion unique identifiers, and even if we exhaust them all, we will reset back to int.MinValue and start over.
It also helps getting rid of locking in the finalizer.